OBMFULL-5463 Can update config on the fly

+++ java/sync/opush/push-dao-ehcache/src/main/java/org/obm/push/store/ehcache/EhCacheConfiguration.java
@@ -114,12 +114,11 @@ public interface EhCacheConfiguration {
 			return percentage;
 		}
 		
-		public int applyTo(int to) {
+		public long applyTo(long to) {
 			Preconditions.checkState(isDefined(), "undefined percentage");
 			Preconditions.checkArgument(to >= 0, "must be a positive value");
-			return Math.round((to / 100f) * percentage);
+			return Math.round((to / 100d) * percentage);
 		}
-
 		
 		@Override
 		public final int hashCode(){
